Fix check-style issues
