+++ presto-hive/src/main/java/io/prestosql/plugin/hive/HiveMetadata.java
@@ -754,8 +754,7 @@ public class HiveMetadata
                             hiveColumnHandle.getHiveColumnIndex(),
                             REGULAR,
                             hiveColumnHandle.getComment(),
-                            Optional.empty())
-                    );
+                            Optional.empty()));
         }
         return columnHandles.build();
     }
